+++ src/components/pages/Tasks/View/index.js
@@ -16,9 +16,14 @@ import {
 } from "apollo-cms/lib/DataView/List/Table";
 import { withStyles } from 'material-ui';
 
-// import TaskStatus from "./Task/Status";
+import TaskStatusSelect from "./Task/Status/Select";
 
 import moment from "moment";
+import { IconButton } from 'material-ui';
+
+import StartIcon from "material-ui-icons/PlayArrow";
+import StopIcon from "material-ui-icons/Stop";
+import { TaskView } from './Task';
 
 
 export class TasksView extends TableView {
@@ -27,27 +32,217 @@ export class TasksView extends TableView {
     ...TableView.propTypes,
     filters: PropTypes.object,
     setFilters: PropTypes.func,
+    createTaskProcessor: PropTypes.func.isRequired,
+    updateTaskProcessor: PropTypes.func.isRequired,
+    createTimerProcessor: PropTypes.func.isRequired,
+    updateTimerProcessor: PropTypes.func.isRequired,
   };
+
+  static defaultProps = {
+    ...TableView.defaultProps,
+    title: "",
+  }
+
+
+
+  getButtons(object) {
+
+    let buttons = [];
+
+
+    const {
+      createTimerProcessor,
+      updateTimerProcessor,
+      classes,
+    } = this.props;
+
+    const {
+      UserLink,
+      //   TaskLink,
+      //   Editor,
+      //   ProjectLink,
+    } = this.context;
+
+
+    if (object) {
+
+
+      const {
+        id: taskId,
+        name,
+        Timers,
+        CreatedBy,
+
+      } = object;
+
+      const {
+        id: createdById,
+      } = CreatedBy || {};
+
+      const {
+        user: currentUser,
+      } = this.context;
+
+      let activeTimers = Timers && Timers.filter(n => n.stopedAt === null) || []
+
+
+      if (activeTimers.length) {
+
+        activeTimers.map(n => {
+          const {
+            id,
+            CreatedBy,
+          } = n;
+
+          buttons.push(<UserLink
+            key={id}
+            user={CreatedBy}
+            size="small"
+            showName={false}
+          />);
+        });
+
+      }
+
+
+      if (currentUser) {
+
+        const {
+          id: currentUserId,
+        } = currentUser;
+
+        const activeTimer = activeTimers.find(n => n.CreatedBy.id === currentUserId);
+
+        if (activeTimer) {
+
+          const {
+            id: timerId,
+          } = activeTimer;
+
+          buttons.push(<IconButton
+            key="stop"
+            onClick={() => updateTimerProcessor({
+              variables: {
+                data: {
+                  stopedAt: new Date(),
+                },
+                where: {
+                  id: timerId,
+                },
+              },
+            })}
+            className={classes.button}
+          >
+            <StopIcon />
+          </IconButton>);
+        }
+        else {
+          buttons.push(<IconButton
+            key="start"
+            onClick={() => createTimerProcessor({
+              variables: {
+                data: {
+                  Task: {
+                    connect: {
+                      id: taskId,
+                    },
+                  },
+                },
+              },
+            })}
+            className={classes.button}
+          >
+            <StartIcon />
+          </IconButton>);
+        }
+
+
+      }
+    }
+
+    return buttons;
+  }
 
 
   getColumns() {
 
 
     const {
+      Grid,
       ProjectLink,
       UserLink,
       TaskLink,
       TaskStatus,
     } = this.context;
 
+    const {
+      classes,
+      updateTaskProcessor,
+      createTimerProcessor,
+      updateTimerProcessor,
+    } = this.props;
+
     return [
+      // {
+      //   id: "Project",
+      //   label: "Проект",
+      //   renderer: (value, record) => {
+      //     return value ? <ProjectLink
+      //       object={value}
+      //     /> : null;
+      //   },
+      // },
+      // {
+      //   id: "id",
+      //   label: "Задача",
+      //   renderer: (value, record) => {
+
+      //     return record ? <TaskLink
+      //       object={record}
+      //     /> : null;
+      //   },
+      // },
       {
-        id: "Project",
-        label: "Проект",
+        id: "status",
+        label: "Статус",
         renderer: (value, record) => {
-          return value ? <ProjectLink
-            object={value}
-          /> : null;
+          // console.log("status", record);
+
+          return value ? <Grid
+            container
+            spacing={8}
+            alignItems="center"
+          >
+            <Grid
+              item
+            >
+              <TaskStatus
+                value={value}
+              />
+              {/* <TaskStatusSelect
+                value={value}
+                inEditMode={true}
+                onChange={event => {
+                  // const {
+                  //   value,
+                  //   name,
+                  // } = event.target;
+                  // this.updateObject({
+                  //   [name]: value,
+                  // });
+                }}
+              /> */}
+            </Grid>
+            {/* <Grid
+              item
+              style={{
+                display: "flex",
+                alignItems: "center",
+              }}
+            >
+              {this.getButtons(record)}
+            </Grid> */}
+          </Grid> : null;
         },
       },
       {
@@ -55,8 +250,16 @@ export class TasksView extends TableView {
         label: "Задача",
         renderer: (value, record) => {
 
-          return record ? <TaskLink
-            object={record}
+          return record ? <TaskView
+            data={{
+              object: record,
+            }}
+            classes={classes}
+            mutate={updateTaskProcessor}
+            createTimer={createTimerProcessor}
+            updateTimer={updateTimerProcessor}
+            showStatus={false}
+            showCreatedBy={false}
           /> : null;
         },
       },
@@ -79,21 +282,10 @@ export class TasksView extends TableView {
         },
       },
       {
-        id: "status",
-        label: "Статус",
-        renderer: (value, record) => {
-          console.log("status", record);
-
-          return value ? <TaskStatus
-            value={value}
-          /> : null;
-        },
-      },
-      {
         id: "RelatedTo",
         label: "Связано с",
         renderer: (value, record) => {
-          console.log("status", TaskLink);
+          // console.log("status", TaskLink);
 
           // return value ? <TaskStatus
           //   value={value}