+++ src/components/pages/Timers/index.js
@@ -59,6 +59,60 @@ class TimersPage extends Page {
     super.componentWillMount && super.componentWillMount();
   }
 
+  
+
+  setFilters(filters) {
+
+    const {
+      uri,
+      router: {
+        history,
+      },
+    } = this.context;
+
+    // console.log("setFilters", filters);
+
+    let newUri = uri.clone();
+
+    try {
+
+      filters = filters ? JSON.stringify(filters) : undefined;
+    }
+    catch (error) {
+      console.error(error);
+    }
+
+    if (filters) {
+
+      if (newUri.hasQuery) {
+        newUri = newUri.setQuery({
+          filters,
+        });
+      }
+      else {
+        newUri = newUri.addQuery({
+          filters,
+        });
+      }
+
+    }
+    else {
+
+      newUri.removeQuery("filters");
+
+    }
+
+    newUri.removeQuery("page");
+
+
+    const url = newUri.resource();
+
+    // console.log("setFilters uri", newUri, url);
+
+    history.push(url);
+
+  }
+
 
   render() {
 
@@ -68,7 +122,7 @@ class TimersPage extends Page {
 
     let {
       first,
-      where,
+      where: propsWhere,
       View,
       ...other
     } = this.props;
@@ -80,8 +134,33 @@ class TimersPage extends Page {
 
     let {
       page,
+      filters,
     } = uri.query(true);
 
+
+    try {
+      filters = filters && JSON.parse(filters) || null;
+    }
+    catch (error) {
+      console.error(console.error(error));
+    }
+
+
+    let AND = [];
+
+    if (propsWhere) {
+      AND.push(propsWhere);
+    }
+
+
+    if (filters) {
+      AND.push(filters);
+    }
+
+
+    let where = {
+      AND,
+    }
 
 
     let skip;
@@ -98,6 +177,8 @@ class TimersPage extends Page {
         first={first}
         skip={skip}
         page={page ? parseInt(page) : undefined}
+        filters={filters || {}}
+        setFilters={filters => this.setFilters(filters)}
         {...other}
       />
     );